Moved interface into interface package.
